+++ src/components/Profile/MyPosts/MyPosts.jsx
@@ -1,0 +1,21 @@
+import s from './MyPosts.module.css'
+import Post from "./Post/Post";
+
+const MyPosts = () => {
+    return (
+        <div className={'main__posts posts'}>
+            <h2 className={'posts__title'}>MyPosts</h2>
+            <textarea className={'posts__input'} placeholder={'Введіть текст...'} cols="20"
+                      rows="3"/>
+            <button className={'posts__button'}>Відправити</button>
+            <Post/>
+            <Post/>
+            <Post/>
+            <Post/>
+            <Post/>
+        </div>
+
+    )
+}
+
+export default MyPosts
