+++ src/components/Profile/MyPosts/MyPosts.module.css
@@ -1,0 +1,2 @@
+/*===============MyPosts====================*/
+
