+++ src/components/Profile/Profile.module.css
@@ -9,9 +9,3 @@
     max-width: 1100px;
 }
 
-.post__img {
-    background-color: darkgray;
-    width: 50px;
-    height: 50px;
-    border-radius: 50%;
-}
